+++ openapi.config.ts
@@ -1,6 +1,6 @@
 import { ConfigFile } from "@rtk-query/codegen-openapi";
 
-const blackList = ["docRetrieve", "authRefreshCreate"];
+const blackList = ["docRetrieve", "authRefreshCreate", "mediaPresignedCreate"];
 
 const openapiConfig: ConfigFile = {
   schemaFile: "http://54.180.123.137/doc/?lang=ko",
@@ -8,9 +8,13 @@ const openapiConfig: ConfigFile = {
   apiImport: "baseApi",
   outputFile: "store/api/injected.ts",
   exportName: "injectedApi",
-  filterEndpoints: (operationName) => {
-    return !blackList.includes(operationName);
+  filterEndpoints: (operationName, operationDefinition) => {
+    // partial update (PATCH)의 open api spec이 update (PUT)과 동일함...
+    return (
+      !blackList.includes(operationName) && operationDefinition.verb !== "patch"
+    );
   },
+
 };
 
 export default openapiConfig;
